+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- add `.file` operation for `trackers` (OSIDB-2029)
+
 ### Fixed
 - request format changed from form-data to json which
   prevents OSIDB serialization failures (OSIDB-2844)